requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
attrs==25.1.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.36.7
botocore==1.36.7
browser-use==0.1.29
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
courlan==1.3.2
dataclasses-json==0.6.7
dateparser==1.2.0
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
filetype==1.2.0
fireworks-ai==0.15.12
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.1
langchain==0.3.16
langchain-anthropic==0.3.3
langchain-aws==0.2.11
langchain-community==0.3.16
langchain-core==0.3.32
langchain-fireworks==0.2.6
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langsmith==0.2.11
lmnr==0.4.55
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
marshmallow==3.26.0
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
ollama==0.4.7
openai==1.60.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-langchain==0.37.1
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-threading==0.50b0
opentelemetry-instrumentation-urllib3==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-semantic-conventions-ai==0.4.2
opentelemetry-util-http==0.50b0
orjson==3.10.15
packaging==24.2
pillow==11.1.0
playwright==1.49.1
posthog==3.10.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.0.0
pyparsing==3.2.1
PySide6==6.8.1.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qasync==0.27.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.11.2
setuptools==75.8.0
shiboken6==6.8.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
tenacity==9.0.0
tiktoken==0.8.0
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
faiss-cpu==1.7.4
sentence-transformers==2.5.1
cross-encoder==0.2.2
scikit-learn==1.4.1.post1
